readme: further details about FWTS, LUV and contributing
People asked more details. Even if this repo is not meant to be a guide, I added 
the minimum information required to run LUV and FWTS. Their documentation
is good enough to not to have to do it ourselves. And running their live image
is really really (really) easy.

Close #1