new: fwtr.sh shell script to help adding new results in the repository.
The first few results made me realize that the logs may contain enough data
to compute the target of results directory in the repository.

The BIOS Information table is used and contains names of vendors.
Note that this table can be very badly programmed by vendors and may require
the user to force values. Hence an interactive command to further validate
the script. Further validation is still required. The script will be
interactive and verbose until we get more confident about the solution.
The BIOS Info parser has only been implemented for FWTS, both html and text results.
The tool (FWTS or LUV so far) prober is implemented for FWTS and LUV. It extracts
which tool has been used, the version and the date the test was run.

To be continued with a 'run' sub-command?
Feel free to enhance :)

Usage example:
```
./scripts/fwtr.sh add my-results-1/results.log my-results-2/results.log my-results-3/results.log
```

See the help message to get the list of options to force some values.
```
./scripts/fwtr.sh --help
```